Must have correct userId.

+++ bugtracker.server/server/services/NoteService.js
@@ -1,10 +1,17 @@
 import { dbContext } from '../db/DbContext'
 
-// import { BadRequest } from '../utils/Errors'
+import { BadRequest } from '../utils/Errors'
 
 class NoteService {
-  async deleteNote(noteId) {
-    return await dbContext.Notes.findByIdAndDelete(noteId)
+  async deleteNote(noteId, userId) {
+    const exists = await dbContext.Notes.findById(noteId)
+    if (!exists) {
+      throw new BadRequest('This Note does not exist')
+    // @ts-ignore
+    } else if (exists._doc.creatorId === userId) {
+      await dbContext.Notes.findByIdAndDelete(noteId)
+      return 'Comment has been delorted!'
+    }
   }
 
   async getNotesByBug(bugId) {
@@ -23,8 +30,15 @@ class NoteService {
     return await dbContext.Notes.create(body)
   }
 
-  async editNote(id, body) {
-    return await dbContext.Notes.findByIdAndUpdate(id, body)
+  async editNote(id, body, userId) {
+    const exists = await dbContext.Notes.findById(id)
+    if (!exists) {
+      throw new BadRequest('This Note does not exist')
+    // @ts-ignore
+    } else if (exists._doc.creatorId === userId) {
+      await dbContext.Notes.findByIdAndUpdate(id, body)
+      return 'Note has been updated!'
+    }
   }
 }
